ENH: rename tree_opt to forest_opt
